rename all files to kebab case
