+++ transpiler/__tests__/end-to-end/mocks/setup/outputs/nest.module.mock.ts
@@ -21,8 +21,8 @@ import { Module, Provider } from '@nestjs/common';
 import {
   PubSubCommandHandlers,
   StreamingCommandHandlers,
-} from './application/commandHandlers/index';
-import { QueryHandlers } from './application/queryHandlers/index';
+} from './application/command-handlers/index';
+import { QueryHandlers } from './application/query-handlers/index';
 import { StreamingDomainEventHandlers } from './application/handlers/domain/index';
 import { StreamingIntegrationEventHandlers } from './application/handlers/integration/index';
